move error proxy to it's own file
